Added new prior Guide method. Renamed and merged previous files together i.e. score -> vector field
