+++ sbi/inference/potentials/vector_field_potential.py
@@ -9,7 +9,10 @@ from torch.distributions import Distribution
 from zuko.distributions import NormalizingFlow
 
 from sbi.inference.potentials.base_potential import BasePotential
-from sbi.inference.potentials.score_fn_util import get_guidance_method, get_iid_method
+from sbi.inference.potentials.vector_field_adaptor import (
+    get_guidance_method,
+    get_iid_method,
+)
 from sbi.neural_nets.estimators import ConditionalVectorFieldEstimator
 from sbi.neural_nets.estimators.shape_handling import (
     reshape_to_batch_event,
